chore(docker): Fix docker entrypoint and - Major refact of the code - Factorize use of API versioning - Update deps
