+++ api/ecoindex/models/examples.py
@@ -1,0 +1,10 @@
+example_daily_limit_response = {
+    "description": "You have reached the daily limit",
+    "content": {
+        "application/json": {
+            "example": {
+                "detail": "You have already reached the daily limit of 5 requests for host www.ecoindex.fr today"
+            }
+        }
+    },
+}
