+++ api/models/enums.py
@@ -1,0 +1,9 @@
+from enum import Enum
+
+
+class Version(Enum):
+    v0 = "v0"
+    v1 = "v1"
+
+    def get_version_number(self):
+        return self.value[1:]
