+++ docker/entrypoint.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+uvicorn api.main:app --host 0.0.0.0 --port 8000
