+++ dockerfile
@@ -10,7 +10,10 @@ COPY --from=requirements-stage /tmp/requirements.txt /code/requirements.txt
 RUN pip install --no-cache-dir --upgrade -r /code/requirements.txt
 RUN pip install psycopg2
 COPY ./ /code/
+COPY ./docker/entrypoint.sh /usr/bin/entrypoint
+RUN chmod +x /usr/bin/entrypoint
 RUN apt-get update \
     && wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb \
     && apt-get install -y ./google-chrome-stable_current_amd64.deb \
     && rm google-chrome-stable_current_amd64.deb
+ENTRYPOINT [ "/usr/bin/entrypoint" ]
